refactor(launcher): move Arrow head bounding into _format
`_bootstrap.py` describes itself as the IPython extension and
installation coordinator, and had grown an Arrow serialization subsystem
to 888 lines. `_format.py` already owns Arrow serialization and
independently decodable chunking, so the mechanics belong there.

`_head_rows_within_bytes`, `_slice_arrow_chunk`, `_collect_arrow_chunks`,
and `_bounded_arrow_table` move to `_format.py` and take their limits as
arguments. `_format.py` reads none of the repr constants, so the split is
policy in bootstrap and mechanism in format. Bootstrap keeps
`_emit_arrow_stream` and the bundle and formatter coordination that is
its job, and drops to 774 lines.

`_summary_hints` was private to `_progressive.py` and imported across a
module boundary. It is now `build_arrow_stream_summary` in `_format.py`,
alongside the manifest builders that use it.

The three repr constants read `NTERACT_ARROW_REPR_*`, matching the
launcher's existing `NTERACT_` flag rather than the `dx` prefix of a
product that no longer ships. They are new in this branch so nothing
depends on the old names. `DX_MAX_PAYLOAD_BYTES` keeps its name because
deployments already set it. All three are documented in the launcher
README.

The module header described what the dx wheel required and why an old
route became unnecessary. It now states the current invariant: no
`ipython_display_formatter` handlers, and both display seats invoking the
same buffer-hook chain.

Behavior is unchanged. PerceptionBench still emits 100 of 3,000 rows at
40.6 MB with `complete: false`, skinny numerics and text still clamp to
50,000 rows, and a 50-row table is still complete.
